refactor: move settings into a own folder
Will start cleaning this up a lot.
