+++ src/lib/parser/Settings/types.ts
@@ -1,0 +1,8 @@
+export interface TemplateFile {
+  parent: string;
+  name: string;
+  front: string;
+  back: string;
+  styling: string;
+  storageKey: string;
+}
